agent: Use bridge iface_name and mac from AgentDB
Replace the use of bridge iface_name and mac on the Agent threads to the one from the AgentDB.
Remove duplicate on backhaul_params on son_slave.
Remove duplicate on m_sConfig on the backhaul manager.
Removing of the flags from the tlvf will take place on the cleanup phase.

PPM-216

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>