agent: Use freq_type and max_bandwidth from AgentDB
"freq_type" and "max_bandwidth" are being sent from the sin_slave to the backhaul
manager which saves a local copy of it.
Replace their use to be from the AgentDB instead of a local copy.

The local backhaul manager copies yet not been removed since at this point we are
still dependent on other members on "slave_sockets".

PPM-216

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>