agent: Use bssids (vaps) list from AgentDB
The bssids (vaps) list is being sent from the son_slave to the backhaul manager when updated,
and the backhaul manager saves a copy of it.
Replace the use of the backhaul manager's local copy of it with the list on the AgentDB.

On this opportunity, refactor the function "send_slave_ap_metric_query_message" to be more
efficient, by not require a mandatory bssid list, and be able to send the query on all bssids
if the given bssid list is empty.

PPM-216

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>