prplmesh-hostapd: added for_all_ap_vaps() with user iterator
There is a need to go on both the ap-vaps configured in hostapd
and in paralell go over the given configuration (non-hostapd structure)
The added functionality gives hostapd configuration class the
responsibility of iterating over the vaps _and_ increment the
user-iterator. the result is that the user's function is called
with the next vap _and_ the next element in its given list.
it is the user responsibility however to make sure that its container
has at least as the number of vaps. otherwise for_all_ap_vaps() would
increment the iterator passed the end of the user's container
tested by implementing the non iterator version of for_all_ap_vaps()
in term of the iterator version, with a dummy iterator

Signed-off-by: Ran Regev <ran.regev@devalore.com>